refactored the interface with the DB
